+++ pkgs/checks/lib/src/extensions/async.dart
@@ -16,13 +16,16 @@ extension FutureChecks<T> on Subject<Future<T>> {
   ///
   /// Fails if the future completes as an error.
   Future<Subject<T>> completes() =>
-      context.nestAsync<T>('completes to a value', (actual) async {
+      context.nestAsync<T>(() => ['completes to a value'], (actual) async {
         try {
           return Extracted.value(await actual);
-        } catch (e) {
-          return Extracted.rejection(
-              actual: ['a future that completes as an error'],
-              which: prefixFirst('threw ', literal(e)));
+        } catch (e, st) {
+          return Extracted.rejection(actual: [
+            'a future that completes as an error'
+          ], which: [
+            ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
+            ...(const LineSplitter()).convert(st.toString())
+          ]);
         }
       });
 
@@ -58,7 +61,7 @@ extension FutureChecks<T> on Subject<Future<T>> {
   ///
   /// Fails if the future completes to a value.
   Future<Subject<E>> throws<E extends Object>() => context.nestAsync<E>(
-          'completes to an error${E == Object ? '' : ' of type $E'}',
+          () => ['completes to an error${E == Object ? '' : ' of type $E'}'],
           (actual) async {
         try {
           return Extracted.rejection(
@@ -66,10 +69,13 @@ extension FutureChecks<T> on Subject<Future<T>> {
               which: ['did not throw']);
         } on E catch (e) {
           return Extracted.value(e);
-        } catch (e) {
+        } catch (e, st) {
           return Extracted.rejection(
               actual: prefixFirst('completed to error ', literal(e)),
-              which: ['is not an $E']);
+              which: [
+                'threw an exception that is not a $E at:',
+                ...(const LineSplitter()).convert(st.toString())
+              ]);
         }
       });
 }
@@ -104,7 +110,7 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
   /// Fails if the stream emits an error instead of a value, or closes without
   /// emitting a value.
   Future<Subject<T>> emits() =>
-      context.nestAsync<T>('emits a value', (actual) async {
+      context.nestAsync<T>(() => ['emits a value'], (actual) async {
         if (!await actual.hasNext) {
           return Extracted.rejection(
               actual: ['a stream'],
@@ -113,10 +119,13 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
         try {
           await actual.peek;
           return Extracted.value(await actual.next);
-        } catch (e) {
+        } catch (e, st) {
           return Extracted.rejection(
               actual: prefixFirst('a stream with error ', literal(e)),
-              which: ['emitted an error instead of a value']);
+              which: [
+                'emitted an error instead of a value at:',
+                ...(const LineSplitter()).convert(st.toString())
+              ]);
         }
       });
 
@@ -131,8 +140,8 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
   /// If this expectation fails, the source queue will be left in it's original
   /// state.
   /// If this expectation succeeds, consumes the error event.
-  Future<Subject<E>> emitsError<E extends Object>() =>
-      context.nestAsync('emits an error${E == Object ? '' : ' of type $E'}',
+  Future<Subject<E>> emitsError<E extends Object>() => context.nestAsync(
+          () => ['emits an error${E == Object ? '' : ' of type $E'}'],
           (actual) async {
         if (!await actual.hasNext) {
           return Extracted.rejection(
@@ -147,10 +156,13 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
         } on E catch (e) {
           await actual.next.then<void>((_) {}, onError: (_) {});
           return Extracted.value(e);
-        } catch (e) {
+        } catch (e, st) {
           return Extracted.rejection(
               actual: prefixFirst('a stream with error ', literal(e)),
-              which: ['emitted an error with an incorrect type, is not $E']);
+              which: [
+                'emitted an error which is not $E at:',
+                ...(const LineSplitter()).convert(st.toString())
+              ]);
         }
       });
 
@@ -193,7 +205,7 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
   /// conditions.
   ///
   /// ```dart
-  /// await checkThat(StreamQueue(someStream)).inOrder([
+  /// await check(StreamQueue(someStream)).inOrder([
   ///   it()..emits().that(it()..equals(0)),
   ///   it()..emits().that(it()..equals(1)),
   //  ]);
@@ -433,16 +445,16 @@ extension ChainAsync<T> on Future<Subject<T>> {
   /// expression that would need parenthesis.
   ///
   /// ```dart
-  /// await checkThat(someFuture).completes().which(it()..equals('expected'));
+  /// await check(someFuture).completes().which(it()..equals('expected'));
   /// // or, with the intermediate `await`:
-  /// (await checkThat(someFuture).completes()).equals('expected');
+  /// (await check(someFuture).completes()).equals('expected');
   /// ```
   Future<void> which(Condition<T> condition) async {
     await condition.applyAsync(await this);
   }
 }
 
-extension StreamQueueWrap<T> on Subject<Stream<T>> {
+extension WithQueueExtension<T> on Subject<Stream<T>> {
   /// Wrap the stream in a [StreamQueue] to allow using checks from
   /// [StreamChecks].
   ///
@@ -450,6 +462,6 @@ extension StreamQueueWrap<T> on Subject<Stream<T>> {
   /// so that they can support conditional expectations and check multiple
   /// possibilities from the same point in the stream.
   Subject<StreamQueue<T>> get withQueue =>
-      context.nest('', (actual) => Extracted.value(StreamQueue(actual)),
+      context.nest(() => [], (actual) => Extracted.value(StreamQueue(actual)),
           atSameLevel: true);
 }