+++ pkgs/checks/test/extensions/async_test.dart
@@ -16,20 +16,20 @@ void main() {
   group('FutureChecks', () {
     group('completes', () {
       test('succeeds for a future that completes to a value', () async {
-        await checkThat(_futureSuccess()).completes().which(it()..equals(42));
+        await check(_futureSuccess()).completes().which(it()..equals(42));
       });
       test('rejects futures which complete as errors', () async {
-        await checkThat(_futureFail()).isRejectedByAsync(
+        await check(_futureFail()).isRejectedByAsync(
           it()..completes().which(it()..equals(1)),
           actual: ['a future that completes as an error'],
-          which: ['threw <UnimplementedError>'],
+          which: ['threw <UnimplementedError> at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<Future<void>>()..completes())
+        await check(it<Future<void>>()..completes())
             .asyncDescription
             .which(it()..deepEquals(['  completes to a value']));
-        await checkThat(it<Future<int>>()..completes().which(it()..equals(42)))
+        await check(it<Future<int>>()..completes().which(it()..equals(42)))
             .asyncDescription
             .which(it()
               ..deepEquals([
@@ -43,12 +43,12 @@ void main() {
       test(
           'succeeds for a future that compeletes to an error of the expected type',
           () async {
-        await checkThat(_futureFail())
+        await check(_futureFail())
             .throws<UnimplementedError>()
             .which(it()..has((p0) => p0.message, 'message').isNull());
       });
       test('fails for futures that complete to a value', () async {
-        await checkThat(_futureSuccess()).isRejectedByAsync(
+        await check(_futureSuccess()).isRejectedByAsync(
           it()..throws(),
           actual: ['completed to <42>'],
           which: ['did not throw'],
@@ -56,17 +56,20 @@ void main() {
       });
       test('failes for futures that complete to an error of the wrong type',
           () async {
-        await checkThat(_futureFail()).isRejectedByAsync(
+        await check(_futureFail()).isRejectedByAsync(
           it()..throws<StateError>(),
           actual: ['completed to error <UnimplementedError>'],
-          which: ['is not an StateError'],
+          which: [
+            'threw an exception that is not a StateError at:',
+            'fake trace'
+          ],
         );
       });
       test('can be described', () async {
-        await checkThat(it<Future<void>>()..throws())
+        await check(it<Future<void>>()..throws())
             .asyncDescription
             .which(it()..deepEquals(['  completes to an error']));
-        await checkThat(it<Future<void>>()..throws<StateError>())
+        await check(it<Future<void>>()..throws<StateError>())
             .asyncDescription
             .which(it()
               ..deepEquals(['  completes to an error of type StateError']));
@@ -75,19 +78,19 @@ void main() {
 
     group('doesNotComplete', () {
       test('succeeds for a Future that never completes', () async {
-        checkThat(Completer<void>().future).doesNotComplete();
+        check(Completer<void>().future).doesNotComplete();
       });
       test('fails for a Future that completes as a value', () async {
         Object? testFailure;
         runZonedGuarded(() {
           final completer = Completer<String>();
-          checkThat(completer.future).doesNotComplete();
+          check(completer.future).doesNotComplete();
           completer.complete('value');
         }, (e, st) {
           testFailure = e;
         });
         await pumpEventQueue();
-        checkThat(testFailure)
+        check(testFailure)
             .isA<TestFailure>()
             .has((f) => f.message, 'message')
             .isNotNull()
@@ -100,13 +103,13 @@ Actual: a future that completed to 'value\'''');
         Object? testFailure;
         runZonedGuarded(() {
           final completer = Completer<String>();
-          checkThat(completer.future).doesNotComplete();
+          check(completer.future).doesNotComplete();
           completer.completeError('error', StackTrace.fromString('fake trace'));
         }, (e, st) {
           testFailure = e;
         });
         await pumpEventQueue();
-        checkThat(testFailure)
+        check(testFailure)
             .isA<TestFailure>()
             .has((f) => f.message, 'message')
             .isNotNull()
@@ -118,7 +121,7 @@ Which: threw 'error'
 fake trace''');
       });
       test('can be described', () async {
-        await checkThat(it<Future<void>>()..doesNotComplete())
+        await check(it<Future<void>>()..doesNotComplete())
             .asyncDescription
             .which(it()..deepEquals(['  does not complete']));
       });
@@ -128,27 +131,27 @@ fake trace''');
   group('StreamChecks', () {
     group('emits', () {
       test('succeeds for a stream that emits a value', () async {
-        await checkThat(_countingStream(5)).emits().which(it()..equals(0));
+        await check(_countingStream(5)).emits().which(it()..equals(0));
       });
       test('fails for a stream that closes without emitting', () async {
-        await checkThat(_countingStream(0)).isRejectedByAsync(
+        await check(_countingStream(0)).isRejectedByAsync(
           it()..emits(),
           actual: ['a stream'],
           which: ['closed without emitting enough values'],
         );
       });
       test('fails for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0)).isRejectedByAsync(
+        await check(_countingStream(1, errorAt: 0)).isRejectedByAsync(
           it()..emits(),
           actual: ['a stream with error <UnimplementedError: Error at 1>'],
-          which: ['emitted an error instead of a value'],
+          which: ['emitted an error instead of a value at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<void>>()..emits())
+        await check(it<StreamQueue<void>>()..emits())
             .asyncDescription
             .which(it()..deepEquals(['  emits a value']));
-        await checkThat(it<StreamQueue<int>>()..emits().which(it()..equals(42)))
+        await check(it<StreamQueue<int>>()..emits().which(it()..equals(42)))
             .asyncDescription
             .which(it()
               ..deepEquals([
@@ -159,25 +162,25 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(queue, it()..emits());
-        await checkThat(queue).emitsError();
+        await check(queue).emitsError();
       });
     });
 
     group('emitsError', () {
       test('succeeds for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0))
+        await check(_countingStream(1, errorAt: 0))
             .emitsError<UnimplementedError>();
       });
       test('fails for a stream that closes without emitting an error',
           () async {
-        await checkThat(_countingStream(0)).isRejectedByAsync(
+        await check(_countingStream(0)).isRejectedByAsync(
           it()..emitsError(),
           actual: ['a stream'],
           which: ['closed without emitting an expected error'],
         );
       });
       test('fails for a stream that emits value', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(
+        await check(_countingStream(1)).isRejectedByAsync(
           it()..emitsError(),
           actual: ['a stream emitting value <0>'],
           which: ['closed without emitting an error'],
@@ -185,20 +188,20 @@ fake trace''');
       });
       test('fails for a stream that emits an error of the incorrect type',
           () async {
-        await checkThat(_countingStream(1, errorAt: 0)).isRejectedByAsync(
+        await check(_countingStream(1, errorAt: 0)).isRejectedByAsync(
           it()..emitsError<StateError>(),
           actual: ['a stream with error <UnimplementedError: Error at 1>'],
-          which: ['emitted an error with an incorrect type, is not StateError'],
+          which: ['emitted an error which is not StateError at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<void>>()..emitsError())
+        await check(it<StreamQueue<void>>()..emitsError())
             .asyncDescription
             .which(it()..deepEquals(['  emits an error']));
-        await checkThat(it<StreamQueue<void>>()..emitsError<StateError>())
+        await check(it<StreamQueue<void>>()..emitsError<StateError>())
             .asyncDescription
             .which(it()..deepEquals(['  emits an error of type StateError']));
-        await checkThat(it<StreamQueue<void>>()
+        await check(it<StreamQueue<void>>()
               ..emitsError<StateError>()
                   .which(it()..has((e) => e.message, 'message').equals('foo')))
             .asyncDescription
@@ -212,25 +215,25 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1);
         await softCheckAsync<StreamQueue<int>>(queue, it()..emitsError());
-        await checkThat(queue).emits().which((it()..equals(0)));
+        await check(queue).emits().which((it()..equals(0)));
       });
     });
 
     group('emitsThrough', () {
       test('succeeds for a stream that eventuall emits a matching value',
           () async {
-        await checkThat(_countingStream(5)).emitsThrough(it()..equals(4));
+        await check(_countingStream(5)).emitsThrough(it()..equals(4));
       });
       test('fails for a stream that closes without emitting a matching value',
           () async {
-        await checkThat(_countingStream(4)).isRejectedByAsync(
+        await check(_countingStream(4)).isRejectedByAsync(
           it()..emitsThrough(it()..equals(5)),
           actual: ['a stream'],
           which: ['ended after emitting 4 elements with none matching'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<int>>()..emitsThrough(it()..equals(42)))
+        await check(it<StreamQueue<int>>()..emitsThrough(it()..equals(42)))
             .asyncDescription
             .which(it()
               ..deepEquals([
@@ -242,25 +245,25 @@ fake trace''');
         final queue = _countingStream(1);
         await softCheckAsync(
             queue, it<StreamQueue<int>>()..emitsThrough(it()..equals(42)));
-        checkThat(queue).emits().which(it()..equals(0));
+        check(queue).emits().which(it()..equals(0));
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
-        await checkThat(queue).emitsThrough(it()..equals(1));
-        await checkThat(queue).emits().which((it()..equals(2)));
+        await check(queue).emitsThrough(it()..equals(1));
+        await check(queue).emits().which((it()..equals(2)));
       });
     });
 
     group('emitsInOrder', () {
       test('succeeds for happy case', () async {
-        await checkThat(_countingStream(2)).inOrder([
+        await check(_countingStream(2)).inOrder([
           it()..emits().which(it()..equals(0)),
           it()..emits().which((it()..equals(1))),
           it()..isDone(),
         ]);
       });
       test('reports which condition failed', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(
+        await check(_countingStream(1)).isRejectedByAsync(
           it()..inOrder([it()..emits(), it()..emits()]),
           actual: ['a stream'],
           which: [
@@ -271,7 +274,7 @@ fake trace''');
         );
       });
       test('nestes the report for deep failures', () async {
-        await checkThat(_countingStream(2)).isRejectedByAsync(
+        await check(_countingStream(2)).isRejectedByAsync(
           it()..inOrder([it()..emits(), it()..emits().which(it()..equals(2))]),
           actual: ['a stream'],
           which: [
@@ -285,7 +288,7 @@ fake trace''');
         );
       });
       test('gets described with the number of conditions', () async {
-        await checkThat(it<StreamQueue<int>>()..inOrder([it(), it()]))
+        await check(it<StreamQueue<int>>()..inOrder([it(), it()]))
             .asyncDescription
             .which(it()..deepEquals(['  satisfies 2 conditions in order']));
       });
@@ -299,7 +302,7 @@ fake trace''');
                 it()..emits().which(it()..equals(1)),
                 it()..emits().which(it()..equals(42)),
               ]));
-        await checkThat(queue).inOrder([
+        await check(queue).inOrder([
           it()..emits().which(it()..equals(0)),
           it()..emits().which(it()..equals(1)),
           it()..emits().which(it()..equals(2)),
@@ -308,8 +311,8 @@ fake trace''');
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
-        await checkThat(queue).inOrder([it()..emits(), it()..emits()]);
-        await checkThat(queue).emits().which(it()..equals(2));
+        await check(queue).inOrder([it()..emits(), it()..emits()]);
+        await check(queue).emits().which(it()..equals(2));
       });
     });
 
@@ -317,17 +320,17 @@ fake trace''');
       test(
           'succeeds for a stream that closes without emitting a matching value',
           () async {
-        await checkThat(_countingStream(5)).neverEmits(it()..equals(5));
+        await check(_countingStream(5)).neverEmits(it()..equals(5));
       });
       test('fails for a stream that emits a matching value', () async {
-        await checkThat(_countingStream(6)).isRejectedByAsync(
+        await check(_countingStream(6)).isRejectedByAsync(
           it()..neverEmits(it()..equals(5)),
           actual: ['a stream'],
           which: ['emitted <5>', 'following 5 other items'],
         );
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<int>>()..neverEmits(it()..equals(42)))
+        await check(it<StreamQueue<int>>()..neverEmits(it()..equals(42)))
             .asyncDescription
             .which(it()
               ..deepEquals([
@@ -339,7 +342,7 @@ fake trace''');
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..neverEmits(it()..equals(1)));
-        await checkThat(queue).inOrder([
+        await check(queue).inOrder([
           it()..emits().which(it()..equals(0)),
           it()..emits().which(it()..equals(1)),
           it()..isDone(),
@@ -349,32 +352,31 @@ fake trace''');
 
     group('mayEmit', () {
       test('succeeds for a stream that emits a matching value', () async {
-        await checkThat(_countingStream(1)).mayEmit(it()..equals(0));
+        await check(_countingStream(1)).mayEmit(it()..equals(0));
       });
       test('succeeds for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0))
-            .mayEmit(it()..equals(0));
+        await check(_countingStream(1, errorAt: 0)).mayEmit(it()..equals(0));
       });
       test('succeeds for a stream that closes', () async {
-        await checkThat(_countingStream(0)).mayEmit(it()..equals(42));
+        await check(_countingStream(0)).mayEmit(it()..equals(42));
       });
       test('consumes a matching event', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(0)));
-        await checkThat(queue).emits().which(it()..equals(1));
+        await check(queue).emits().which(it()..equals(1));
       });
       test('does not consume a non-matching event', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(1)));
-        await checkThat(queue).emits().which(it()..equals(0));
+        await check(queue).emits().which(it()..equals(0));
       });
       test('does not consume an error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(0)));
-        await checkThat(queue)
+        await check(queue)
             .emitsError<UnimplementedError>()
             .which(it()..has((e) => e.message, 'message').equals('Error at 1'));
       });
@@ -382,32 +384,32 @@ fake trace''');
 
     group('mayEmitMultiple', () {
       test('succeeds for a stream that emits a matching value', () async {
-        await checkThat(_countingStream(1)).mayEmitMultiple(it()..equals(0));
+        await check(_countingStream(1)).mayEmitMultiple(it()..equals(0));
       });
       test('succeeds for a stream that emits an error', () async {
-        await checkThat(_countingStream(1, errorAt: 0))
+        await check(_countingStream(1, errorAt: 0))
             .mayEmitMultiple(it()..equals(0));
       });
       test('succeeds for a stream that closes', () async {
-        await checkThat(_countingStream(0)).mayEmitMultiple(it()..equals(42));
+        await check(_countingStream(0)).mayEmitMultiple(it()..equals(42));
       });
       test('consumes matching events', () async {
         final queue = _countingStream(3);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..isLessThan(2)));
-        await checkThat(queue).emits().which(it()..equals(2));
+        await check(queue).emits().which(it()..equals(2));
       });
       test('consumes no events if no events match', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..isLessThan(0)));
-        await checkThat(queue).emits().which(it()..equals(0));
+        await check(queue).emits().which(it()..equals(0));
       });
       test('does not consume an error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..equals(0)));
-        await checkThat(queue)
+        await check(queue)
             .emitsError<UnimplementedError>()
             .which(it()..has((e) => e.message, 'message').equals('Error at 1'));
       });
@@ -415,16 +417,16 @@ fake trace''');
 
     group('isDone', () {
       test('succeeds for an empty stream', () async {
-        await checkThat(_countingStream(0)).isDone();
+        await check(_countingStream(0)).isDone();
       });
       test('fails for a stream that emits a value', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(it()..isDone(),
+        await check(_countingStream(1)).isRejectedByAsync(it()..isDone(),
             actual: ['a stream'], which: ['emitted an unexpected value: <0>']);
       });
       test('fails for a stream that emits an error', () async {
         final controller = StreamController<void>();
         controller.addError('sad', StackTrace.fromString('fake trace'));
-        await checkThat(StreamQueue(controller.stream)).isRejectedByAsync(
+        await check(StreamQueue(controller.stream)).isRejectedByAsync(
             it()..isDone(),
             actual: ['a stream'],
             which: ['emitted an unexpected error: \'sad\'', 'fake trace']);
@@ -432,10 +434,10 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1);
         await softCheckAsync<StreamQueue<int>>(queue, it()..isDone());
-        await checkThat(queue).emits().which(it()..equals(0));
+        await check(queue).emits().which(it()..equals(0));
       });
       test('can be described', () async {
-        await checkThat(it<StreamQueue<int>>()..isDone())
+        await check(it<StreamQueue<int>>()..isDone())
             .asyncDescription
             .which(it()..deepEquals(['  is done']));
       });
@@ -443,13 +445,13 @@ fake trace''');
 
     group('emitsAnyOf', () {
       test('succeeds for a stream that matches one condition', () async {
-        await checkThat(_countingStream(1)).anyOf([
+        await check(_countingStream(1)).anyOf([
           it()..emits().which(it()..equals(42)),
           it()..emits().which((it()..equals(0)))
         ]);
       });
       test('fails for a stream that matches no conditions', () async {
-        await checkThat(_countingStream(0)).isRejectedByAsync(
+        await check(_countingStream(0)).isRejectedByAsync(
             it()
               ..anyOf([
                 it()..emits(),
@@ -467,7 +469,7 @@ fake trace''');
             ]);
       });
       test('includes nested details for nested failures', () async {
-        await checkThat(_countingStream(1)).isRejectedByAsync(
+        await check(_countingStream(1)).isRejectedByAsync(
             it()
               ..anyOf([
                 it()..emits().which(it()..equals(42)),
@@ -487,7 +489,7 @@ fake trace''');
             ]);
       });
       test('gets described with the number of conditions', () async {
-        await checkThat(
+        await check(
                 it<StreamQueue<int>>()..anyOf([it()..emits(), it()..emits()]))
             .asyncDescription
             .which(it()..deepEquals(['  satisfies any of 2 conditions']));
@@ -501,40 +503,44 @@ fake trace''');
                 it()..emits().which(it()..equals(10)),
                 it()..emitsThrough(it()..equals(42)),
               ]));
-        await checkThat(queue).emits().which(it()..equals(0));
+        await check(queue).emits().which(it()..equals(0));
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
-        await checkThat(queue).anyOf([
+        await check(queue).anyOf([
           it()..emits().which(it()..equals(1)),
           it()..emitsThrough(it()..equals(1))
         ]);
-        await checkThat(queue).emits().which(it()..equals(2));
+        await check(queue).emits().which(it()..equals(2));
       });
     });
   });
 
   group('ChainAsync', () {
     test('which', () async {
-      await checkThat(_futureSuccess()).completes().which(it()..equals(42));
+      await check(_futureSuccess()).completes().which(it()..equals(42));
     });
   });
 
   group('StreamQueueWrap', () {
     test('can wrap streams in a queue', () async {
-      await checkThat(Stream.value(1)).withQueue.emits();
+      await check(Stream.value(1)).withQueue.emits();
     });
   });
 }
 
 Future<int> _futureSuccess() => Future.microtask(() => 42);
 
-Future<int> _futureFail() => Future.error(UnimplementedError());
+Future<int> _futureFail() =>
+    Future.error(UnimplementedError(), StackTrace.fromString('fake trace'));
 
 StreamQueue<int> _countingStream(int count, {int? errorAt}) => StreamQueue(
       Stream.fromIterable(
         Iterable<int>.generate(count, (index) {
-          if (index == errorAt) throw UnimplementedError('Error at $count');
+          if (index == errorAt) {
+            Error.throwWithStackTrace(UnimplementedError('Error at $count'),
+                StackTrace.fromString('fake trace'));
+          }
           return index;
         }),
       ),