+++ pkgs/checks/CHANGELOG.md
@@ -1,6 +1,12 @@
-# 0.1.1-dev
+# 0.2.0
 
-- Added an example.
+-   **Breaking Changes**
+    -   `checkThat` renamed to `check`.
+    -   `nest` and `nestAsync` take `Iterable<String> Function()` arguments for
+        `label` instead of `String`.
+-   Added an example.
+-   Include a stack trace in the failure description for unexpected errors from
+    Futures or Streams.
 
 # 0.1.0